* ipa-inline.h (ipa_call_summary): Turn sizes into signed; 	add ctor. 	* ipa-inline.c (want_inline_small_function_p): Do not cast to 	unsigned.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@248263 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/ipa-inline.c
@@ -723,7 +723,7 @@ want_inline_small_function_p (struct cgraph_edge *e, bool report)
 	   && (!e->count || !e->maybe_hot_p ()))
 	   && inline_summaries->get (callee)->min_size
 		- ipa_call_summaries->get (e)->call_stmt_size
-	      > (unsigned)MAX (MAX_INLINE_INSNS_SINGLE, MAX_INLINE_INSNS_AUTO))
+	      > MAX (MAX_INLINE_INSNS_SINGLE, MAX_INLINE_INSNS_AUTO))
     {
       e->inline_failed = CIF_MAX_INLINE_INSNS_AUTO_LIMIT;
       want_inline = false;
@@ -731,7 +731,7 @@ want_inline_small_function_p (struct cgraph_edge *e, bool report)
   else if ((DECL_DECLARED_INLINE_P (callee->decl) || e->count)
 	   && inline_summaries->get (callee)->min_size
 		- ipa_call_summaries->get (e)->call_stmt_size
-	      > (unsigned)16 * MAX_INLINE_INSNS_SINGLE)
+	      > 16 * MAX_INLINE_INSNS_SINGLE)
     {
       e->inline_failed = (DECL_DECLARED_INLINE_P (callee->decl)
 			  ? CIF_MAX_INLINE_INSNS_SINGLE_LIMIT
